add debug_config method to server to print config to CLI
