+++ counterpartylib/lib/backend/addrindex.py
@@ -18,16 +18,6 @@ raw_transactions_cache = util.DictCache(size=config.BACKEND_RAW_TRANSACTIONS_CAC
 unconfirmed_transactions_cache = None
 reverse_unconfirmed_transactions_cache = None
 
-URL_USERNAMEPASS_REGEX = re.compile('.+://(.+)@')
-
-
-def clean_url_for_log(url):
-    m = URL_USERNAMEPASS_REGEX.match(url)
-    if m and m.group(1):
-        url = url.replace(m.group(1), 'XXXXXXXX')
-
-    return url
-
 
 class BackendRPCError(Exception):
     pass
@@ -46,7 +36,7 @@ def rpc_call(payload):
                 logger.debug('Successfully connected.')
             break
         except (Timeout, ReadTimeout, ConnectionError):
-            logger.debug('Could not connect to backend at `{}`. (Try {}/{})'.format(clean_url_for_log(url), i+1, TRIES))
+            logger.debug('Could not connect to backend at `{}`. (Try {}/{})'.format(util.clean_url_for_log(url), i+1, TRIES))
             time.sleep(5)
 
     if response == None:
@@ -54,7 +44,7 @@ def rpc_call(payload):
             network = 'testnet'
         else:
             network = 'mainnet'
-        raise BackendRPCError('Cannot communicate with backend at `{}`. (server is set to run on {}, is backend?)'.format(clean_url_for_log(url), network))
+        raise BackendRPCError('Cannot communicate with backend at `{}`. (server is set to run on {}, is backend?)'.format(util.clean_url_for_log(url), network))
     elif response.status_code not in (200, 500):
         raise BackendRPCError(str(response.status_code) + ' ' + response.reason)
