+++ counterpartylib/lib/backend/btcd.py
@@ -4,7 +4,7 @@ import sys
 import json
 
 from counterpartylib.lib import script
-from counterpartylib.lib import config
+from counterpartylib.lib import config, util
 
 import requests
 from requests.exceptions import Timeout, ReadTimeout, ConnectionError
@@ -34,7 +34,7 @@ def rpc_call(payload):
                 logger.debug('Successfully connected.')
             break
         except (Timeout, ReadTimeout, ConnectionError):
-            logger.debug('Could not connect to backend at `{}`. (Try {}/{})'.format(url, i+1, TRIES))
+            logger.debug('Could not connect to backend at `{}`. (Try {}/{})'.format(util.clean_url_for_log(url), i+1, TRIES))
             time.sleep(5)
 
     if response == None:
@@ -42,7 +42,7 @@ def rpc_call(payload):
             network = 'testnet'
         else:
             network = 'mainnet'
-        raise BackendRPCError('Cannot communicate with backend at `{}`. (server is set to run on {}, is backend?)'.format(url, network))
+        raise BackendRPCError('Cannot communicate with backend at `{}`. (server is set to run on {}, is backend?)'.format(util.clean_url_for_log(url), network))
     elif response.status_code not in (200, 500):
         raise BackendRPCError(str(response.status_code) + ' ' + response.reason)
